--- src/app/auth/login/page.tsx
@@ -1,13 +1,0 @@
-import { LoginForm } from "./page.client";
-
-const Page = () => {
-  return (
-    <div className="flex justify-center w-full">
-      <div className="w-full max-w-sm">
-        <LoginForm />
-      </div>
-    </div>
-  );
-};
-
-export default Page;
